requirements.txt
asgiref==3.10.0
Django==5.2.7
django-bootstrap5==25.2
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
